fix freeze on removing module (#2)
* fix freeze on removing module

When removing the module it checks for the dma_virt_addr and if it
is not NULL it tries to free the dma allocation. However since
dma_virt_addr is never initialized, it holds a garbage value which
ends up causing a freeze on the system.

* Add missing initializations.

length initialization is missing on the regions[0] for some devices.
There is also minor comma addition.

* fix wrongfully incremented obj counter in rare cases.

If device_create errors out, then the obj counter is already
incremented. Instead moved the incrementation after the error
check.