Add flag to check for RH builds
* Add check for flag for RH builds

Red Hat has a a history of backporting changes that break interfaces.
In the mainline kernel these changes are handled via LINUX_VERSION_CODE
macro, but RH will backport them anyway.
This commit adds the logic to check for the flag, and changes the logic
of the #if to make it more readable going forward.